requirements.txt
fastapi
uvicorn[standard]
python-socketio[asgi]
python-dotenv
aiofiles
pymongo
motor
asyncpg
python-jose[cryptography]
google-genai
google-generativeai
google-adk
starlette
httpx
redis[hiredis]
Pillow
python-multipart
boto3
